requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django==5.0.6
django_rest_framework==0.1.0
# openai==1.30.5
langchain==0.2.1
langchain-community==0.2.1
langchainhub==0.1.17
langchain-openai==0.1.8
# bs4==0.0.2
opensearch-py==2.6.
boto3==1.34.117
pandas==2.1.4
openpyxl==3.1.2
